+++ wayvr/Cargo.toml
@@ -85,6 +85,7 @@ tracing = "0.1.43"
 tracing-subscriber = { version = "0.3.22", features = ["env-filter"] }
 uuid = { workspace = true }
 wayland-client = { workspace = true }
+wayland-protocols-misc = { workspace = true }
 winit = { version = "0.30.12", optional = true }
 xcb = { version = "1.6.0", features = [
   "as-raw-xcb-connection",
