+++ wayvr/src/subsystem/hid/provider/dummy.rs
@@ -1,0 +1,16 @@
+use glam::Vec2;
+use crate::subsystem::hid::{VirtualKey, WheelDelta};
+use crate::subsystem::hid::provider::HidProvider;
+
+pub struct DummyProvider;
+
+impl HidProvider for DummyProvider {
+    fn mouse_move(&mut self, _pos: Vec2) {}
+    fn send_button(&mut self, _button: u16, _down: bool) {}
+    fn wheel(&mut self, _delta: WheelDelta) {}
+    fn set_modifiers(&mut self, _modifiers: u8) {}
+    fn send_key(&self, _key: VirtualKey, _down: bool) {}
+    fn set_desktop_extent(&mut self, _extent: Vec2) {}
+    fn set_desktop_origin(&mut self, _origin: Vec2) {}
+    fn commit(&mut self) {}
+}
